[heap] Remove ExternalBackingStoreType::kArrayBuffer
This CL removes ExternalBackingStoreType::kArrayBuffer. We only
use the counter for array buffers in
NewSpace::ExternalBackingStoreOverallBytes(). This method is used by
Heap::AllocateExternalBackingStoreBytes().

We keep this working as-is by adding a new method
YoungAllocatedExternalBackingStoreBytes() for this instead. It
combines NewSpace::ExternalBackingStoreOverallBytes() and the
young array buffer size.

This CL does not address the bytes_ overflow issue in the linked
bug yet. This will be done in a follow-up CL.

Bug: 460166671
Change-Id: I30f4a2b0cebd8aed95c743dfa72d9611a8239b68
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7159853
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#103787}