[strings] Prevent OOM crash in URI encoding
The std::vector used as a buffer in EncodeURI can fail to allocate when
encoding very large strings. This ends up OOMing and crashing the
process.

This patch replaces the std::vector with a custom ResizableBuffer that
handles allocation failures gracefully. This allows the builtin to stop
processing and throw a InvalidStringLengthError instead of crashing.

Since this is a new failure mode, additionally do some cleanup of the
encode method to use a non-allocating helper which returns a status.

Bug: 459950607
Change-Id: I8d580832297b2c16c8b651c2f691e916c56d6390
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7156424
Auto-Submit: Leszek Swirski <leszeks@chromium.org>
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Commit-Queue: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#103788}