[turboshaft] Embed AssemblerOpInterface into reducer stack
The Assembler no longer inherits directly from both, the ReducerStack
and the AssemblerOpInterface, but the latter is part of the former's
class hierarchy. This allows reducers to overwrite assembler functions
(e.g. builtin reducers can provide a CallBuiltin that doesn't require
a FrameState argument).

Furthermore: Remove the additional layer of TS-abstraction in the
Assembler setup that we introduced when we tried to make the
Turboshaft architecture IR independent.

NO_IFTTT=Only changed class name inside a comment.

Bug: 348031042
Change-Id: If8b9b27a9655102b7c85353a8511482297c7c451
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7167247
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Commit-Queue: Nico Hartmann <nicohartmann@chromium.org>
Cr-Commit-Position: refs/heads/main@{#103808}