[heap] No need to finish sweeping in final atomic pause
We were trying to finish sweeping even if incremental marking was
already running. While this is technically correct because we check
whether sweeping is in progress for each component we want to
finalize first, conceptually we shouldn't finish sweeping in such
situations. It can be a bit confusing in a trace because
EnsureSweepingCompleted() is unconditionally emitting a trace event. In
addition we would also call UpdateAllocationLimits(), which emits
its own trace event as well.

Bug: 456384548
Change-Id: Ia5d1322a5599c1bfb09b7e15df9b7a3fb3653df6
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7172159
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#103805}