[heap] Allow overflow of bytes_ in ArrayBufferList::Append()
We were DCHECKing that this overflow doesn't happen but it is
possible on 32-bit when e.g. growing a buffer in small steps.
In such situations it can happen that we have plenty of
ArrayBufferExtensions for each version of that buffer that are still
accounted but actually already detached and their memory freed.

Since this shouldn't happen in practice and we also only use that
counter for GC scheduling purposes, we can allow this overflow here
by removing the DCHECK.

Bug: 460166671
Change-Id: I15b2d9e7554e49dcefa8a2a430b92c383c76199f
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7170480
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#103812}