[heap] Move marking start tracing event
Move V8.GCIncrementalMarkingStart to the beginning of
Heap::StartIncrementalMarking(). This way this event also covers
the safepoint and finishing sweeping.

For IncrementalMarking::Start() we have another tracing event, so we
don't lose information for it. But we didn't have one for
Heap::StartIncrementalMarking() so far.

Bug: 456384548
Change-Id: Ia46c03d27adb437946aa3c11bbd67254bef6ad9a
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7172483
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#103827}