[wasm] Add isolate-specific info to AsyncCompileJob later
This CL adds an `AsyncCompileJob::IsolateSpecificInfo` struct which is
only populated later (after initialization, potentially after passing
bytes to the `AsyncCompileJob`, but before finishing the async
compilation).

This will be used for the new API for source phase imports where the
isolate, context etc are only known later in the pipeline, after the
embedder finished (streaming) background compilation from a foreground
task.

The main change is that counter updates and the UKM metrics event for
decoding need to be delayed until we get hold of an isolate.

R=jkummerow@chromium.org

Bug: 446885278
Change-Id: I29683a5fd404d6112d1f2daa3ca4b68608eaed15
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7157255
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#103859}