[maglev] Support the construct_as_builtin path when inlining ctors
This case goes through the JSBuiltinsConstructStub builtin (vs.
JSConstructStubGeneric). Contrary to the latter, the receiver is always
the_hole, and the return value is assumed to be a JSReceiver. In Maglev,
we inline both of these stubs.

Note that we have to be careful around the_hole receivers, as this
case is not particularly well supported in CallKnownJSFunction and
friends. We already had this case before (for derived constructors),
but it looks like we implicitly avoided problematic paths for this
case.

Change-Id: I02e0df01bf428f2e181ff87d1f5e2b4130460f72
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7165876
Commit-Queue: Jakob Linke <jgruber@chromium.org>
Reviewed-by: Victor Gomes <victorgomes@chromium.org>
Cr-Commit-Position: refs/heads/main@{#103845}