[*lev] Keep DataView alive as long as it's being used
A recent refactoring (https://crrev.com/c/7141782) of DataView in
Maglev allowed reusing the data pointer between subsequent DataView
loads/stores.

However, this lead to removing references to the DataView object
itself, which allows the GC to free it. This CL ensures that despite
this new optimization, we still keep a reference to the DataView for
as long as we're accessing it.

Change-Id: I8105f08a9a9005d8d69a5d15ed5b7503a3ad3992
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7184739
Reviewed-by: Marja Hölttä <marja@chromium.org>
Commit-Queue: Darius Mercadier <dmercadier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#103871}