[wasm][d8] Fix concurrent manipulation during deserialization
The wire bytes and serialized bytes both could potentially be
manipulated while we try to deserialize them.
Avoid the problem by copying both before deserialization.
I also switch from internal objects to using the public API, which
provides easier to use methods for the copy.

Note that this only affects d8.

The wire bytes will be copied again in `DeserializeNativeModule`, which
looks like an unnecessary copy in most cases (both in d8 and in the
streaming decoder we have an owned copy of the wire bytes already).
I will fix this in a follow-up CL.

R=jkummerow@chromium.org

Bug: 461547336
Change-Id: Ic3ace4eae4a5849407d23cc8860a74fa5052489a
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7184738
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#103912}