[wasm][ESM] Add API for source-phase imports
The API is similar to the existing `WasmStreaming` API, but can be
instantiated fully from the background. Only the final `Finish` step
needs to be called from a foreground task.

After all the refactorings that already landed for
https://crbug.com/446885278 this is a pretty straight-forward addition
now which reuses most of the existing logic for `WasmStreaming`.

R=jkummerow@chromium.org
CC=lpardosixtos@microsoft.com

Bug: 446885278
Change-Id: I4749e27a722b1e50e107a808a3e02f8d70adbbce
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7156414
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#103915}