[compiler] Add %MajorGCForCompilerTesting intrinsic
So far, the way to invoke GCs from JS was to use `gc()` in combination
with --expose-gc. However, this was leading to GC calls being generic
runtime calls in the optimizing compilers, which was blocking all kind
of optimizations. As such, weird tricks had to be used in order to
repro GC timing issues, cf the repros in https://crrev.com/c/7184739
(which this CL nicely simplifies).

The new %MajorGCForCompilerTesting intrinsic is marked as no-read 
no-write in Turbofan, Turboshaft and Maglev, which means that it 
won't block any optimizations and will allow to more easily fuzz 
and repro GC-timing-related bugs.

Change-Id: I66cb03c77f350880b2667ca5ae44ab823ddb8f34
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7198340
Reviewed-by: Nico Hartmann <nicohartmann@chromium.org>
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Darius Mercadier <dmercadier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#103914}