[wasmfx] Implement tag parameters and returns
Pass the tag arguments and returns via a stack buffer. The buffer is
immediately unpacked by the receiving stack and never used again, so we
don't need to visit the tagged slots within that buffer.
The pointer to the buffer is passed as an extra argument to the
stack switching builtins, and forwarded to the target stack.
On the receiving end:
- The suspend builtin returns the pointer received from the resume
  builtin,
- The resume handler uses a new pseudo-instruction to materialize the
  buffer in a fixed register, similar to how we materialize the
  exception or the continuation object.

Arguments and return values in the initial func ref are not supported
yet.

R=jkummerow@chromium.org

Bug: 388533754
Change-Id: I0618b5f41e3ac9981c7d5e2c1214f3ec87300757
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7170116
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Thibaud Michaud <thibaudm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#103917}