[wasm] Add stream state checks
Track the state of the streaming decoder, and explicitly disallow calls
to `Finish` after `Finish` or `Abort` have already been called.

Do allow `Finish` after `NotifyCompilationDiscarded` though, as that's a
situation which can happen and which is fine (the user is navigating
away anyway, so it does not matter if we still resolve the promise or
not).

R=jkummerow@chromium.org

Bug: 462888125
Change-Id: I6d98185d8b82768165665e1510675bbb15639189
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7198047
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#103934}