[wasm interpreter] Add trace in LoadMem/StoreMem builtin
When use LoadMem/StoreMem builtin, the trace execution in LoadMem
and StoreMem will get a wrong result. Because the TracePop and
TracePush haven't been call in these builtin. And the trace execution
is important to fix a bug, so I fix it in this CL.

Fixed: 461128968
Change-Id: I4e8da9ed3591f1711e5aa6c216593453796a4481
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7157156
Reviewed-by: Daniel Lehmann <dlehmann@chromium.org>
Commit-Queue: Paolo Severini <paolosev@microsoft.com>
Reviewed-by: Paolo Severini <paolosev@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#103941}