[riscv][wasmfx] Implement tag parameters and returns
Port commit 3d95b453b017ea741a8f6a81dcf5ace31ff5140a
Bug: 388533754

Origin message:
Pass the tag arguments and returns via a stack buffer. The buffer is
immediately unpacked by the receiving stack and never used again, so we
don't need to visit the tagged slots within that buffer.
The pointer to the buffer is passed as an extra argument to the
stack switching builtins, and forwarded to the target stack.
On the receiving end:
- The suspend builtin returns the pointer received from the resume
  builtin,
- The resume handler uses a new pseudo-instruction to materialize the
  buffer in a fixed register, similar to how we materialize the
  exception or the continuation object.

Arguments and return values in the initial func ref are not supported
yet.

Change-Id: I367cf07f72217fc3fa7a5894e6822371f780b56e
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7202304
Commit-Queue: Ji Qiu <qiuji@iscas.ac.cn>
Reviewed-by: Ji Qiu <qiuji@iscas.ac.cn>
Cr-Commit-Position: refs/heads/main@{#103925}