[wasm][turboshaft] Fix Wasm-in-JS inlining of void functions
Creating a JS undefined as a result of a Wasm function call with zero
return values is already done by the wrapper, so the body inlining
should simply produce an invalid value.

To make the possible result states a bit more clear, I introduce a new
tagged union for the result of the body inlining: failed, successful
but with void result, successful with single return value.

Note that the mixed wrapper inlining in the Turbofan frontend already
took care to never read this result of the call (and instead produces a
JS undefined for each use), so this unnecessary value was never
consumed.

The new wrapper inlining in Turbolev/Turboshaft also correctly ignored
this result.

Fixed: 462045796
Change-Id: Iee5b50376559f9914da658f548733b4b44904e8a
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7177861
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Commit-Queue: Daniel Lehmann <dlehmann@chromium.org>
Cr-Commit-Position: refs/heads/main@{#103949}