[sandbox] Expand trusted object unpublishing mechanism
This CL expands the "publishing" mechanism for trusted objects.
Previously, we already had a special "unpublished" tag that could be
used to make trusted objects inaccessible from within the sandbox. We
now expand the machinery around this tag to allow objects to be
allocated in an unpublished state and then be published later on. For
that, we also fix a bug that previously allowed unpublished objects to
still be accessed from inside the sandbox through the
kUnknownIndirectPointerTag (used e.g. for the SFI).

The new, more flexible unpublishing mechanism will be used in follow-up
CLs to make bytecode inaccessible until it has been validated.

Drive-By: rename init_self_indirect_pointer to InitAndPublish since this
method does non-trivial work and so should probably be uppercased.

Bug: 461681036
Change-Id: Iad0224beb89ef1bd625b5cdc51a77cef12642159
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7172412
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#103956}