[sandbox] Introduce AbortWithSandboxViolation
Normally, for sandbox testing/fuzzing, we ignore Abort(), CHECK() etc.
as these represent controlled (and therefore safe) crashes. However,
there are cases where we can detect a compromised state, but only too
late to reliably mitigate exploitation. For these cases, we now
introduce a new AbortWithSandboxViolation function which can be used to
crash but still report a sandbox violation during fuzzing.

One specific use case is the interpreter executing an unknown bytecode.
This typically means that we're executing arbitrary bytecode, and just
happen to execute an invalid opcode. However, as the bytecode may be
attacker-controlled, this is likely an exploitable condition. Prior to
this change, our sandbox fuzzers would not treat this as a sandbox
violation as we'd crash with a nullptr deref. With this CL, we now use
AbortWithSandboxViolation to report this appropriately.

Bug: 461681036
Change-Id: I7afee7bf7c762cae6ffd7b8515c693e8e1401466
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7203123
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#103959}