[flags] Handle --proto-assign-seq-opt-count=0
We shouldn't run the optimization if the value is zero.
Specifying a negative implication from that case to the feature flag
will disable it if it isn't explicitly enabled. If it is, this will
crash the process with a meaningful error message.

This is a follow-up to https://crrev.com/c/7178541.

Bug: 429332174
Change-Id: Ifa9f0720d6593b89b37f5408109327813159f19b
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7204863
Auto-Submit: Matthias Liedtke <mliedtke@chromium.org>
Reviewed-by: Raphael Herouart <rherouart@chromium.org>
Commit-Queue: Raphael Herouart <rherouart@chromium.org>
Cr-Commit-Position: refs/heads/main@{#103979}