[heap] Remove external backing store bytes tracking for space/page
This CL removes tracking of external backing store bytes per page and
per space. kExternalString was the only ExternalBackingStoreType left.
We never used the per-page counter and only used the per-space
counter for YoungAllocatedExternalBackingStoreBytes() but since we do not have young external strings, we can just drop it.

However, we maintain a total counter for external strings in
Heap::backing_store_bytes_ along with ArrayBuffers which are
reported in GetHeapStatistics(). This CL removes that field but
keeps this working by switching the Heap::backing_store_bytes()
method to ExternalStringTable::GetBytes().

External strings are currently not accounted in external memory for
GC scheduling. This CL does not change this but it adds a counter
to the external string table which is currently only used for the
tests in test-external-string-tracker.cc. A follow-up CL will add this
tracking.

Bug: 460166671
Change-Id: I1a08270fc0cf4df3ac98e78ecb08d9df5fbdaf09
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7203286
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#103988}