[wasm] Avoid wire bytes copy on deserialization
The `DeserializeNativeModule` always (when preconditions were met)
created a copy of the passed wire bytes. The caller often already has an
owned copy though, so the additional copy is unnecessary.

This CL thus removes the copy in `DeserializeNativeModule` and takes
over the owned copy from the caller.

What complicates this is that ownership of the owned copy must remain
with the caller when deserialization fails. So we pass a non-const
reference to the owned vector, and ownership is only *conditionally*
transferred (when deserialization succeeded).

R=jkummerow@chromium.org

Change-Id: I21d88c7caa9e3531550ace331e9e44be9cb38b88
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7185804
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#103987}