[v8windbg] Harden Isolate* retrieval in v8windbg
In recent builds v8windbg didn't correctly calculate the Isolate* from
a debugged crash dump. We used `offsetof`` to calculate Isolate* from
Heap*, given that the Heap object is contained in the Isolate object,
but in some builds the offset from the crash dump turned out to be
different from the one compiled in the v8windbg plugin. This caused an
error while trying to read the content of strings.
As a fix, we now use the PDB symbols loaded by WinDbg to exactly
calculate this offset.

Change-Id: I7ea5508c3fae3ce56ae69bf054284f74c502137f
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7205515
Reviewed-by: Daniel Lehmann <dlehmann@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Paolo Severini <paolosev@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#104175}