[sandbox] Build foundation for bytecode verification
This CL creates the foundation for building bytecode verification and
enforcing that all bytecode is verified before use. It does not yet
implement any actual verification, which will be added in future CLs.

Specifically, we introduce the BytecodeVerifier class that will be able
to verify bytecode in the future. Currently the verification is a no-op.
Then, we modify the BytecodeArray creation machinery to initially create
BytecodeArray objects in an unpublished state (see crrev.com/c/7172412)
and only make them accessible to the sandbox after verification. This
helps ensure that all bytecode is verified prior to use.

Bug: 461681036
Change-Id: I024db361f85d4aca24fae9cbd665d29b2b78fd64
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7169863
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#104003}