`Runtime.addBinding` can be limited to main world
Previously, in `Runtime.addBinding` command, the `executionContextName`
with value empty string is considered to be an invalid value, as it used
internally to mark the binding to be global. The side effect was not
being able to add a binding only to the default execution context, and
make it not available to custom ones.

This PR splits the global and per-execution context bindings, which
allows for an empty string to be a valid execution context name
identifier. Which in turn allows for adding binding only for the default
execution context.

Bug: 399595414
Change-Id: I30131db09a3eb6e20041cb0ee8f04474fed64c63
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7209047
Auto-Submit: Maksim Sadym <sadym@chromium.org>
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Reviewed-by: Simon Zünd <szuend@chromium.org>
Cr-Commit-Position: refs/heads/main@{#104015}