[turboshaft] Introduce kAny Select implementation
In order to make it easier to emit CMoves: currently every caller of
Select has to check whether CMove is available and has to think about
"would it be better with a CMove or a Branch?".

With this CL, Select callers can just use Implementation::kAny when
they don't care about the implementation, and let the compiler select
what it think is better.

Note that this CL doesn't change any of the existing Select uses; a
follow-up will relax a few uses to use kAny instead of kBranch.

Change-Id: I67173b22a694608f0db332f67bc2d3b7cfbea8a8
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7213430
Commit-Queue: Jakob Kummerow <jkummerow@chromium.org>
Auto-Submit: Darius Mercadier <dmercadier@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#104044}