[sandbox] Prevent UAF of dispatch handle
The `InstantiateAsmJs` builtin calls out to JS code and afterwards uses
the passed-in dispatch handle for a tail call.
Nothing (except for in-sandbox objects) keeps the dispatch handle alive
though. So the dispatch handle could be freed and re-allocated by the JS
code and the tail-call would call code with a different parameter count,
corrupting the stack.

This CL adds a check for the expected parameter count just before the
tail call.

See the linked issue for a discussion on a more complete mitigation of
the problem.

R=olivf@chromium.org
CC=mlippautz@chromium.org

Bug: 462217236
Change-Id: I3a91b7b70253fd31ed2bb507feeb130e0959df36
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7210825
Reviewed-by: Olivier Flückiger <olivf@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#104059}