[regexp] Fix RegExpBytecodeIterator on big-endian
Because we are packing the first argument with the bytecode (writing
4-byte as uint32_t) when writing RegExp bytecode, we can't simply read
the current byte to identify the bytecode in the RegExpBytecodeIterator
due to endianness issues.
Instead we need to also read 4-byte as uint32_t and extract the
bytecode.

Drive-by: Also fix `--print-regexp-bytecode` on big-endian, which had
the same issue as the iterator.

Bug: 433891213, 437003349
Change-Id: I9b25d6b24bcd817da78a498426d4bc88ab3f3fda
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7215952
Auto-Submit: Patrick Thier <pthier@chromium.org>
Reviewed-by: Jakob Linke <jgruber@chromium.org>
Commit-Queue: Patrick Thier <pthier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#104069}