[sandbox] Add support for fuzzing the BytecodeVerifier
This CL exposes the new BytecodeVerifier to fuzzers so that it can be
tested properly. For that, we introduce two new native functions for
testing: %GetBytecode which returns the BytecodeArray of a given
function as a JSObject and %InstallBytecode which takes such an object,
converts it back to a BytecodeArray, passes it to the verifier, and
finally installs it on the given function. This then allows writing
simple fuzzers to mutate the raw bytecode of an existing function and
thereby test the verifier.

Bug: 461681036
Change-Id: I7fc09490ee998b5adef34bf5bbcd2c379d1b0adb
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7207156
Commit-Queue: Samuel Groß <saelo@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#104087}