[regexp] BC: SkipUntilBitInTable advance_by should be 32-bit
Similar to WriteCurrentPositionToRegister and AdvanceRegister,
advance_by is used as 32-bit in the old bytecode format, but as 16-bit
in the new format.
Change advance_by to Offset32 in the new format for backwards
compatibility.
This fixes an issue with assembling RegExp code from bytecode on
big-endian.

Bug: 433891213, 437003349
Change-Id: I581fb067e26cc5882201d3549ce4e9e6bb56845d
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7222870
Commit-Queue: Patrick Thier <pthier@chromium.org>
Reviewed-by: Jakob Linke <jgruber@chromium.org>
Cr-Commit-Position: refs/heads/main@{#104084}