[compiler] Optimize conversions for inlined JS->Wasm wrappers
Wasm functions which return an i32 (which is probably the most-used
return type) currently unconditionally convert its result to a
HeapNumber / Smi using low-level primitives.

This change replaces that with a high-level JS primitive when inlining
a JS->Wasm wrapper.
This optimizes richards-wasm in JetStream3 containing:
> while (scheduleIter()) {}
(where scheduleIter is a JS function calling a Wasm function which
returns an `i32`.)

After special-casing the conversions that happen in such cases, the
result from the call to the non-inlined wasm function returning a
Word32 register can be directly used as the input to the conditional
jump that implements the loop backedge.

DISABLE_SPELLCHECKER

Bug: 465068841
Change-Id: Iaddc63b24295434c81184b0e900220a70c6656ea
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7213424
Reviewed-by: Daniel Lehmann <dlehmann@chromium.org>
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Commit-Queue: Matthias Liedtke <mliedtke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#104090}