[wasm][turbolev] Avoid inlining duplication in TS wrapper
Currently the Turbolev/Turboshaft JS-to-Wasm wrapper inlining inlines
the body twice, once for the fast path for argument conversion, and once
for the slow path. This change refactors the code to inline only once.

We can probably remove the fast path / slow path split in a followup
completely, but let's do that in a separate CL to attribute regressions
more easily to one or the other. (Well, there should not be any
regressions due to this CL here.)

Bug: 455592073
Change-Id: I925141da80cc33f81d2f40ffe8e0de2692a4f057
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7223870
Reviewed-by: Paolo Severini <paolosev@microsoft.com>
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Commit-Queue: Daniel Lehmann <dlehmann@chromium.org>
Cr-Commit-Position: refs/heads/main@{#104093}