[regexp] Rename FixedLengthLoopState to SpecialLoopState.
While generating the fixed length loops (.*) we have
somewhere to store labels and other state.  This gives
us more flexibility on how to code generate, for example
we set up the counter on the stack on entering the stack
and any backtracking goes to code that checks that counter.

We want this flexibility for the omnivorous non-greedy
loops too.  These are the loops that are prepended to
non-anchored regexps so they can quickly search for
potential match positions.

This is a renaming-only CL. (It also adds one argument
to a function that is not yet used.) The intention is to
make reviewing simpler by separating renames and actual
new functionality, which will arrive in a later CL.

Change-Id: I720761f50c09c6005e3182a9b4b511cd4c639809
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7224414
Commit-Queue: Erik Corry <erikcorry@chromium.org>
Reviewed-by: Patrick Thier <pthier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#104110}