[*lev] Support inlining of async functions
This is actually very easy to support, since inlined async functions
never need to suspend/resume in the inlined body: to suspend, async
functions return a promise, and the callers awaiting this promise is
what really suspends. And the key thing is that when execution
resumes, it will resume in a non-inlined version of the async
function.
This means that inlined async functions never resume, and as a result,
they never lead to edges that bypass loop headers (which is one of the
hard thing when compiling generator functions).

Bug: 431933185
Change-Id: I504fa3cb631d51f5acfeb29a9450e555b93530d2
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7227458
Auto-Submit: Darius Mercadier <dmercadier@chromium.org>
Commit-Queue: Darius Mercadier <dmercadier@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#104129}