[cppgc] Fix CycleEndHistogramReportsCorrectValues test
The test checks that scope durations are as expected. In time of high
cpu contention, the test can get preempted while a scope is open, thus
increasing the duration and causing DCHECKs to fail.
Instead we can just check that the duration is at least the expected
value. It's less strict but shouldn't falkily crash.

Bug: 465398124
Change-Id: I524ca22449309b0a2aafd7b0ee55bd9b1190660c
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7228222
Reviewed-by: Dominik Inführ <dinfuehr@chromium.org>
Commit-Queue: Omer Katz <omerkatz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#104116}