[sandbox] Check that forward jumps jump forward in BytecodeVerifier
With this change, the BytecodeVerifier now also ensures that forward
jumps actually jump forward and, in particular, not back to the current
instruction. This is needed because if we ever end up processing a
forward jump with offset zero, then the baseline compiler (and possibly
other compilers as well) will end up generating invalid machine code.
This is because the jump will initially be emitted as a placeholder jump
(to an invalid offset) and we assume that we'll patch it to become a
valid jump later on. However, for jumps to the current instruction, this
will not happen (see BaselineCompiler::VisitSingleBytecode) and we're
left with invalid code.

Bug: 461681036
Change-Id: I412eb7dd2304f5fcca18eddaa86a838b0b4ba526
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7228662
Commit-Queue: Samuel Groß <saelo@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#104130}