[strings] Fix SharedStringAccessGuard for strings in shared space
Previously SharedStringAccessGuard::GetIsolateIfNeeded() returned the
shared space isolate only for shared strings, but we need to return
it for all strings that reside in shared space. These are not the same.
Strings that can be in-place internalized are automatically allocated in
shared space, even though they are not shared.

Fixed: 465680239
Change-Id: I7414959f66938a1beb785863f2a4394b98fa033f
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7228905
Reviewed-by: Dominik Inführ <dinfuehr@chromium.org>
Commit-Queue: Patrick Thier <pthier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#104132}