[ignition] Invalidate BytecodeArrayIterator when done
Refactor BytecodeArrayIterator to null its cursor when it reaches the
end of the iteration. Any subsequent attempt to advance the iterator,
will end up accessing the zero page, which can safely crash.

As part of this, bottleneck the cursor access in a new
UpdateCurrentBytecode (replacing the old UpdateOperandScale), which
caches the read bytecode on the iterator (rather than re-reading it on
every access). Methods that check that the offset is within a pointer
range have to now check the cursor being in range of the pointer range,
in case it was nulled out by iteration.

Bug: 461681036
Change-Id: I4de61a3dd89f1daf93e2d030a8a5c8b115537da0
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7228948
Commit-Queue: Leszek Swirski <leszeks@chromium.org>
Reviewed-by: Samuel Groß <saelo@chromium.org>
Auto-Submit: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#104137}