[inspector] Don't delete Channel on session disconnect
With https://crrev.com/c/6433946 we keep sessions alive for a bit longer
when we disconnect while paused. This is because agent code could still
be on the stack and do stuff once the stack is torn down.

The test harness currently removes the channel very close to the time
when a session disconnects. This leads to a UAF (in the test harness
ONLY) where the session tries to send CDP things after disconnect.

This CL fixes the UAF by delaying channel cleanup until we tear down the
whole test harness.

Fixed: 466244926
Change-Id: I783b3dc679e513c275abfbd8d1bd3e60b367ae28
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7234284
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>
Cr-Commit-Position: refs/heads/main@{#104159}