[*lev] Mark generator body as dead when inlining it
Cf comment in the code: when inlining a call to a generator function,
we'll only execute the initialization of the generator, since the body
of the function is only executed when calling .next() (which we don't
currently inline).

So far, we were relying on subsequent optimizations to detect that the
body of the generator was dead. With this CL, we now skip emitting
this dead code from the start. This should improve compile time, and
might also help other optimizations whose states' won't be polluted by
dead values anymore.

Fixed: 466333912
Bug: 431933185
Change-Id: If0ab4e5a8952e859a83652528ee29f2cfce706fa
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7231999
Commit-Queue: Darius Mercadier <dmercadier@chromium.org>
Auto-Submit: Darius Mercadier <dmercadier@chromium.org>
Reviewed-by: Victor Gomes <victorgomes@chromium.org>
Cr-Commit-Position: refs/heads/main@{#104165}