[api] Fix contextual accesses within detached context
... broken by https://crrev.com/c/7184979.

The issue was that
 - for "contextual" accesses to JSGlobalObject properties V8 needs to
   provide respective JSGlobalProxy as the "holder" to the
   accessor/interceptor callbacks,
 - the "holder" was computed via JSGlobalObject::global_proxy() which
   in case of detached context might already be backed by another
   instance of JSGlobalObject.
 - as a result if an Api callback would try to access the detached
   global object via such "holder" it'd access wrong JSGlobalObject
   instance - it'd be the current one instead of original one.

The fix is to create a new JSGlobalProxy instance during context
detaching, wire it properly with the old JSGlobalObject instance and
pass it as a "holder" value to Api callbacks. This other global proxy
is stored in a new JSGlobalObject::global_proxy_for_api field.

Thus, the detached global object gets new identity which can be
observed only in accessor/interceptor callbacks as the "holder" will
no longer be equal to any of the `v8::Context::Global()`.

To summarize behavior changes related to Holder()->HolderV2() migration
(avoid exposing JSGlobalObject through Api) for accessor/interceptors
on the global object, previously we had the following:
  CHECK_NE(attached_context->Global(), info.Holder());
  CHECK_NE(detached_context->Global(), info.Holder());

but now we'll have:
  CHECK_EQ(attached_context->Global(), info.HolderV2());
  CHECK_NE(detached_context->Global(), info.HolderV2());

Additionally, this CL adds Realm.navigateSameOrigin(i) for testing.

Fixed: 464961493
Fixed: 463593768
Bug: 333672197
Change-Id: Iee26ddd496252309ad59f43b34f644c411c11986
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7232534
Reviewed-by: Toon Verwaest <verwaest@chromium.org>
Commit-Queue: Igor Sheludko <ishell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#104166}